Move modules to a new sub-package state_machine
