+++ src/nextline_schedule/auto/state_machine/__init__.py
@@ -1,0 +1,5 @@
+__all__ = [
+    'AutoModeStateMachine',
+]
+
+from .machine import AutoModeStateMachine
