+++ src/nextline_schedule/auto/state_machine/factory.py
