+++ src/nextline_schedule/auto/state_machine/machine.py
